sched: add MockClock/MockIrqSource behind sched-mock feature (#668)
Closes #668.

RFC 0005 wave 3: lands the first non-production `Clock` /
`IrqSource` impls behind a new `sched-mock` Cargo feature, plus a
host unit test that exercises them end-to-end (advance the clock,
inject a timer IRQ, observe a wakeup).

`MockClock` and `MockIrqSource` live in `kernel/src/task/env.rs`
gated by `#[cfg(feature = "sched-mock")]` only — no `target_os`
exception — so the nm-check guard (#669) can statically verify the
symbols don't appear in a release-built kernel ELF. The mocks own
all their state internally (no touching `crate::time::WAKEUPS` or
`crate::arch::*`) so test-driven advances cannot collide with the
production PIT tick path.

To make the host unit test reachable, the bare-metal-only contents
of `task/mod.rs` are extracted into a new `task/sched_core.rs`
gated to `target_os = "none"`, while `task::env` widens to
`#[cfg(any(target_os = "none", all(test, feature = "sched-mock")))]`.
Pure mechanical move; no behaviour change in the scheduler core
proper.

Side fix: `task::init`'s production-env identity assertion is moved
from `sched_core` into a new `env::assert_production_env()` helper.
`core::ptr::eq` on `&dyn Trait` compares both data and vtable
pointers, and the vtable for a given trait/type pair is materialized
per codegen unit — once the assertion's two `&'static dyn`
materializations lived in different modules they saw different
vtable pointers and the assertion spuriously fired. Keeping both
materializations inside `env.rs` itself sidesteps the issue.

Out of scope (test-side counterpart to #667's call-site migration):
the `with_mock_env` thread-local override helper and converting an
existing in-kernel scheduler integration test to drive the mocks
both depend on `env()` being on the call path, which #667 lands.
This PR only ships the trait impls and host-side proof-of-life.

Test plan:
- `cargo fmt --all -- --check`
- `cargo xtask build` (no sched-mock; release-shaped surface unchanged)
- `cargo test --lib --features sched-mock task::env::mock_tests`
  (7 tests pass: seed honored, advance saturates, drain order,
  shared-deadline drain, ack/inject counters, end-to-end wakeup)
- `cargo xtask test`
- `cargo xtask smoke`
